build: 날짜관련 유틸함수 분리
